refactor : 파일 이름 및 패키지 변경 (#582)
